+++ v3/v3.go
@@ -1,8 +1,8 @@
 package v3
 
 import (
+	"crypto/x509"
 	"fmt"
-	"strings"
 
 	"github.com/nutanix-cloud-native/prism-go-client"
 	"github.com/nutanix-cloud-native/prism-go-client/internal"
@@ -12,49 +12,53 @@ const (
 	libraryVersion = "v3"
 	absolutePath   = "api/nutanix/" + libraryVersion
 	userAgent      = "nutanix/" + libraryVersion
-	clientName     = "prism_central"
 )
 
 // Client manages the V3 API
 type Client struct {
 	client *internal.Client
 	V3     Service
+
+	clientOpts []internal.ClientOption
+}
+
+// ClientOption is a functional option for the Client
+type ClientOption func(*Client) error
+
+// WithCertificate sets the certificate for the client
+func WithCertificate(certificate *x509.Certificate) ClientOption {
+	return func(c *Client) error {
+		c.clientOpts = append(c.clientOpts, internal.WithCertificate(certificate))
+		return nil
+	}
 }
 
 // NewV3Client return a internal to operate V3 resources
-func NewV3Client(credentials prismgoclient.Credentials) (*Client, error) {
-	var baseClient *internal.Client
-
-	// check if all required fields are present. Else create an empty internal
-	if credentials.Username != "" && credentials.Password != "" && credentials.Endpoint != "" {
-		c, err := internal.NewClient(
-			internal.WithCredentials(&credentials),
-			internal.WithUserAgent(userAgent),
-			internal.WithAbsolutePath(absolutePath))
-		if err != nil {
+func NewV3Client(credentials prismgoclient.Credentials, opts ...ClientOption) (*Client, error) {
+	v3Client := &Client{}
+	for _, opt := range opts {
+		if err := opt(v3Client); err != nil {
 			return nil, err
 		}
-		baseClient = c
-	} else {
-		errorMsg := fmt.Sprintf("Prism Central (PC) Client is missing. "+
-			"Please provide required details - %s in provider configuration.", strings.Join(credentials.RequiredFields[clientName], ", "))
-
-		baseClient = &internal.Client{UserAgent: userAgent, ErrorMsg: errorMsg}
 	}
 
-	f := &Client{
-		client: baseClient,
-		V3: Operations{
-			client: baseClient,
-		},
+	if credentials.Username == "" || credentials.Password == "" || credentials.Endpoint == "" {
+		return nil, fmt.Errorf("username, password and endpoint are required")
 	}
 
-	// f.internal.OnRequestCompleted(func(req *http.Request, resp *http.Response, v interface{}) {
-	// 	if v != nil {
-	// 		utils.PrintToJSON(v, "[Debug] FINISHED REQUEST")
-	// 		// TBD: How to print responses before all requests.
-	// 	}
-	// })
+	v3Client.clientOpts = append(v3Client.clientOpts,
+		internal.WithCredentials(&credentials),
+		internal.WithUserAgent(userAgent),
+		internal.WithAbsolutePath(absolutePath),
+	)
 
-	return f, nil
+	httpClient, err := internal.NewClient(v3Client.clientOpts...)
+	if err != nil {
+		return nil, err
+	}
+
+	v3Client.client = httpClient
+	v3Client.V3 = Operations{client: httpClient}
+
+	return v3Client, nil
 }