+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added the "environment" package which provides abstraction for retrieving settings like API endpoints, credentials and their sources to evolve independently from clients
 - Add logr based configurable logging for internal.NewClient
+- Add `WithCertificate` functional option for v3 client constructor
 
 ### Changed
 - The http client has been moved from pkg/nutanix to repo root
@@ -19,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The root package is renamed from `prism_go_client` to `prismgoclient`
 - Modify NewClient constructor to use functional options
 - NewRequest and NewAuthRequest methods on internal.Client don't admit context in params
+- v3 client constructor now takes functional options as parameters
+- v3 client constructor returns error instead of failing silently
 
 ### Removed
 - remove internal.NewBaseClient constructor